Adds comments, custom-fields to issues list, pass to view func
When using the `issues list` command, in the interactive TUI you can
call the `issue view` command for selected issues. However there was no
way to specify the options to view that issue with. This adds the
`--comments` and `--custom-fields` options to the `issue list` command
and passes those options through to `issue view` as part of the
`ViewOptions` struct.